requirements.txt
streamlit>=1.25.0
google-generativeai>=0.1.0
